requirements.txt
asgiref==3.11.0
Django==5.2.8
djangorestframework==3.16.1
sqlparse==0.5.3
django-cors-headers==4.9.0
